lib: Figure out if we are being starved for cpu
Add two things to lib/thread.c:

a) If a thread timer should have popped CPU_CONSUMED_CHECK
seconds in the past, and we are only handling it now.  Consider
the thread starved and notice it.

b) If the wall time for a running thread minus the cpu time
for a running thread is greater than CPU_CONSUMED_CHECK / 2
seconds consider the thread as having been starved for
cpu.

Signed-off-by: Donald Sharp <sharpd@nvidia.com>