vtysh: Aggregate key chain config from multiple modules
Problem Statement:
=================
Currently RIP and EIGRP register for keychain. so when keychain is
configured, both modules get the config and when we execute
"show running-config" we get two copies of configuration from RIP
and EIGRP. If more modules use keychain then we get more copies of
the config in running-config.

RCA:
====
Since keychain is implemented as lib it runs in the context of the
process using it. Many copies of keychain config get generated due
to this.

Fix:
====
There is similar problem with route-map and other libs that run in
module context.

There is some extra problem with keychain since there is sub mode
in CLI. so uniqueness is specific to the mode the configuration is
appplied.

fix is to aggregate the CLI based on the mode where its configured.

Risk:
=====
Low risk

Tests Executed:
===============
Have tested multiple scenarios with same config in multiple modes.
Without special handling u cannot have cryptographic-algorithm
and key-string having same values in "key chain pqr".

Eg:
!
key chain abcd
 key 100
  key-string password
  cryptographic-algorithm sha1
 exit
 key 200
  key-string password
  cryptographic-algorithm sha256
 exit
!
key chain pqr
 key 300
  key-string password
  cryptographic-algorithm sha384
 exit
 key 400
  key-string password
  cryptographic-algorithm sha384
 exit
!

Signed-off-by: Abhinay Ramesh <rabhinay@vmware.com>